requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
langchain==0.2.6
langchain-groq==0.1.6
pymongo==4.7.3
python-dotenv==1.0.1
pydantic==2.7.4

FYuqBTM8MUQaUGTyLgq2bcaCTp
